requirements.txt
django==1.8.4
djangorestframework==3.2.0
django-rest-swagger==0.3.10
pymysql==0.7.10